requirements.txt
anyio==3.5.0
asgiref==3.5.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
config==0.5.1
dnspython==2.2.1
email-validator==1.2.0
fastapi==0.75.2
h11==0.13.0
httptools==0.4.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
orjson==3.6.8
packaging==21.3
pluggy==1.0.0
py==1.11.0
pydantic==1.9.0
pymongo==4.1.1
pyparsing==3.0.8
pytest==7.1.2
python-dotenv==0.20.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.27.1
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
tomli==2.0.1
typing_extensions==4.2.0
ujson==5.2.0
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
